requirements.txt
streamlit
pandas
google-genai
plotly
mysql-connector-python
python-dotenv